+++ app/src/main/java/com/android/airmart/data/dao/CommentDao.kt
@@ -7,7 +7,7 @@ import com.android.airmart.data.entity.Comment
 @Dao
 interface CommentDao {
     @Query("SELECT * from comment WHERE productId= :productId ORDER BY createdAt")
-    fun getAllCommentsByProductId(productId:Int): LiveData<List<Comment>>
+    fun getAllCommentsByProductId(productId:Long): LiveData<List<Comment>>
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     fun insertComment(comment : Comment):Long
     @Update
